requirements.txt
alabaster==0.7.12
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
black==18.9b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
concrete-settings==0.0.1
docutils==0.14
entrypoints==0.3
factory-boy==2.11.1
Faker==1.0.4
filelock==3.0.10
flake8==3.7.7
flake8-bugbear==18.8.0
idna==2.8
imagesize==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==6.0.0
packaging==19.0
pluggy==0.9.0
pudb==2018.1
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
pyparsing==2.3.1
pytest==4.3.1
pytest-mock==1.10.1
pytest-pudb==0.7.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
text-unidecode==1.2
toml==0.10.0
tox==3.7.0
urllib3==1.24.1
urwid==2.0.1
virtualenv==16.4.3
